scripts/generate_silences_cr/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
idna==3.3
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
